requirements.txt
fastapi
    uvicorn
    python-multipart
    youtube-transcript-api
    aiohttp
    pydantic
    groq
    python-dotenv
